+++ xash3d_engine/engine/ref/soft/r_part.c
@@ -248,7 +248,7 @@ void GAME_EXPORT CL_DrawTracers(double frametime, particle_t* cl_active_tracers)
 			VectorAdd(verts[0], delta, verts[2]);
 			VectorAdd(verts[1], delta, verts[3]);
 
-			if ( p->color > sizeof(gTracerColors) / sizeof(color24) )
+			if ( (size_t)p->color > sizeof(gTracerColors) / sizeof(color24) )
 			{
 				gEngfuncs.Con_Printf(S_ERROR "UserTracer with color > %zu\n", sizeof(gTracerColors) / sizeof(color24));
 				p->color = 0;
